tests: (almost) remove use of "test_i18ngrep !"
As noted in previous commits we are removing uses of
test_i18ngrep. Let's start with the more rare negated form. This
leaves us with "test_i18ngrep" as a mere wrapper for "grep" at the
end.

Done with:

    cd t &&
    perl -pi -e 's/test_i18ngrep ! /! grep /g' *.sh

I excluded those things that would conflict with other in-flight
changes on the "seen" branch, as well as not removing support from
this invocation from "test_i18ngrep" for those & any other in-flight
and soon-to-be-submitted changes. We can do a later final cleanup and
removal of the unused code in some later change.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>